+++ task_10/index.html
@@ -3,8 +3,8 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
-    <script src="homework.js" defer></script>
+    <link rel="stylesheet" href="styles/main.css">
+    <script type="module" src="scripts/main.js" defer></script>
     <title>Библиотека</title>
 </head>
 <body>
